+++ Ecodan.h
@@ -48,7 +48,7 @@ public:
   void SetHolidayMode(uint8_t OnOff);
   void SetSvrControlMode(uint8_t OnOff, uint8_t DHW, uint8_t Z1H, uint8_t Z1C, uint8_t Z2H, uint8_t Z2C);
   void GetFTCVersion(void);
-  void SetHotWaterSetpoint(uint8_t Target);
+  void SetHotWaterSetpoint(float Target);
   void SetHeatingControlMode(uint8_t Mode, uint8_t Zone);
   void SetSystemPowerMode(uint8_t OnOff);
   void WriteMELCloudCMD(uint8_t cmd);
